Show the delete/cancel icon only for the imported node
Imported nodes will not have edit icon on their node tile. Hence, it is required to show the delete icon in their node tile.
As the other nodes will have edit icon and their edit forms contain the delete button, it is not necessary to show the
delete/cancel icon in their node tiles.